image/requirement.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
bcrypt==3.1.6
billiard==3.3.0.23
boto==2.49.0
boto3==1.9.130
botocore==1.12.130
celery==3.1.26.post2
cffi==1.12.2
cryptography==2.6.1
Django==2.2
django-celery==3.2.2
django-storages==1.7.1
djangorestframework==3.9.2
docutils==0.14
jmespath==0.9.4
kombu==3.0.37
paramiko==2.4.2
Pillow==6.0.0
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.0
pytz==2019.1
redis==2.10.3
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
storage==0.0.4.3
urllib3==1.24.1
vine==1.3.0
